Add a NullKernel for C_R to create an invertible matrix
The C_R variable lives on the entire 1-D block but previously it
was only getting residual and Jacobian contributions from the NodalKernel
on the left boundary. This resulted in MOOSE/PETSc not even allocating
diagonal jacobian entries for C_R which leads to segmentation faults
or errors from preconditioners like hpyre's boomeramg. To fix this
we add a NullKernel for C_R which adds small diagonal entries for
dofs in the interior domain